requirements.txt
discord.py
flask[async]
python-dotenv
asyncpg
